老控制板存在bug:1号端口设置50%开度时，会导致故障，修改软件适应老控制板，当设置50%开度时，程序按49.96发送指令
git-svn-id: https://192.168.1.124:8443/svn/delusoft/trunk/qtverify@1508 62fcab80-fd93-7748-8c49-e77283212eb3

+++ source/comobject/source/comobject.cpp
@@ -312,7 +312,7 @@ void ControlComObject::askControlRegulate(UINT8 portno, int degree)
 	}
 	qDebug()<<"ControlComObject::askControlRegulate portno ="<<portno<<", degree ="<<degree;
 	QByteArray buf;
-	m_controlProtocol->makeFrameOfCtrlRegulate(portno, degree);
+	m_controlProtocol->makeFrameOfCtrlRegulate(portno, (float)degree);
 	buf = m_controlProtocol->getSendBuf();
 	int size = buf.size();
 	for (int i=0; i<size; i++)
